+++ models.go
@@ -10,20 +10,33 @@ import (
 type ComputeSize string
 
 const (
-	ComputeSizeXS ComputeSize = "XS"
-	ComputeSizeS  ComputeSize = "S"
-	ComputeSizeM  ComputeSize = "M"
-	ComputeSizeL  ComputeSize = "L"
-	ComputeSizeXL ComputeSize = "XL"
+	ComputeSizeXS   ComputeSize = "XS"
+	ComputeSizeS    ComputeSize = "S"
+	ComputeSizeM    ComputeSize = "M"
+	ComputeSizeL    ComputeSize = "L"
+	ComputeSizeXL   ComputeSize = "XL"
+	ComputeSize2XL  ComputeSize = "2XL"
+	ComputeSize3XL  ComputeSize = "3XL"
+	ComputeSize4XL  ComputeSize = "4XL"
+	ComputeSizeAuto ComputeSize = "AUTO"
 )
 
-type UpsertMode string
+type QueryFormat string
 
 const (
-	UpsertModeCreate    UpsertMode = "create"
-	UpsertModeAppend    UpsertMode = "append"
-	UpsertModeUpsert    UpsertMode = "upsert"
-	UpsertModeOverwrite UpsertMode = "overwrite"
+	QueryFormatDefault QueryFormat = "default"
+	QueryFormatCSV     QueryFormat = "csv"
+	QueryFormatJSONL   QueryFormat = "jsonl"
+	QueryFormatParquet QueryFormat = "parquet"
+)
+
+type UploadMode string
+
+const (
+	UploadModeCreate       UploadMode = "create"
+	UploadModeAppend       UploadMode = "append"
+	UploadModeCreateAppend UploadMode = "create_append"
+	UploadModeOverwrite    UploadMode = "overwrite"
 )
 
 type SessionKind string
@@ -108,7 +121,9 @@ type QueryRequest struct {
 	Cache       *bool        `json:"cache,omitempty"`
 	Catalog     *string      `json:"catalog,omitempty"`
 	ComputeSize *ComputeSize `json:"compute_size,omitempty"`
+	Dialect     *string      `json:"dialect,omitempty"`
 	Ephemeral   *bool        `json:"ephemeral,omitempty"`
+	Format      *QueryFormat `json:"format,omitempty"`
 	Limit       *int         `json:"limit,omitempty"`
 	Offset      *int         `json:"offset,omitempty"`
 	QueryID     *string      `json:"query_id,omitempty"`
@@ -220,11 +235,12 @@ type CancelQueryResponse struct {
 }
 
 type UpsertParams struct {
-	Catalog    string
-	Schema     string
-	Table      string
-	Mode       UpsertMode
-	PrimaryKey string
+	Catalog     string
+	Schema      string
+	Table       string
+	PrimaryKey  string
+	CursorField string
+	ContentType string
 }
 
 func (p UpsertParams) values() url.Values {
@@ -233,11 +249,31 @@ func (p UpsertParams) values() url.Values {
 		"schema":  []string{p.Schema},
 		"table":   []string{p.Table},
 	}
-	if p.Mode != "" {
-		values.Set("mode", string(p.Mode))
-	}
 	if p.PrimaryKey != "" {
 		values.Set("primary_key", p.PrimaryKey)
+	}
+	if p.CursorField != "" {
+		values.Set("cursor_field", p.CursorField)
+	}
+	return values
+}
+
+type UploadParams struct {
+	Catalog     string
+	Schema      string
+	Table       string
+	Mode        UploadMode
+	ContentType string
+}
+
+func (p UploadParams) values() url.Values {
+	values := url.Values{
+		"catalog": []string{p.Catalog},
+		"schema":  []string{p.Schema},
+		"table":   []string{p.Table},
+	}
+	if p.Mode != "" {
+		values.Set("mode", string(p.Mode))
 	}
 	return values
 }